CI: Test Swift
Partially-implements: blueprint kolla-deployment-scenario-testing

+++ tests/setup_disks.sh
@@ -1,13 +1,27 @@
+#!/bin/bash
+
+# $1: scenario / ceph store type
+
+set -o xtrace
+set -o errexit
+
 mkdir -p /opt/data/kolla
 
-if [ $1 = 'cinder-lvm' ]; then
+if [ $1 = 'cinder-lvm' ] || [ $1 = 'zun' ]; then
     # cinder-volumes volume group
     free_device=$(losetup -f)
     fallocate -l 5G /var/lib/cinder_data.img
     losetup $free_device /var/lib/cinder_data.img
     pvcreate $free_device
     vgcreate cinder-volumes $free_device
-
+elif [ $1 = 'swift' ]; then
+    # swift partition
+    free_device=$(losetup -f)
+    fallocate -l 5G /var/lib/swift_data.img
+    losetup $free_device /var/lib/swift_data.img
+    parted $free_device -s -- mklabel gpt mkpart KOLLA_SWIFT_DATA 1 -1
+    free_partition=${free_device}p1
+    mkfs.xfs -L d0 $free_partition
 elif [ $1 = 'filestore' ]; then
     #setup devices for Kolla Ceph filestore OSD
     dd if=/dev/zero of=/opt/data/kolla/ceph-osd1.img bs=5M count=1000
@@ -19,7 +33,7 @@ elif [ $1 = 'filestore' ]; then
     LOOP=$(losetup -f)
     losetup $LOOP /opt/data/kolla/ceph-journal1.img
     parted $LOOP -s -- mklabel gpt mkpart KOLLA_CEPH_OSD_BOOTSTRAP_OSD1_J 1 -1
-else
+elif [ $1 = 'bluestore' ]; then
     # Setup devices for Kolla Ceph bluestore OSD
     dd if=/dev/zero of=/opt/data/kolla/ceph-osd0.img bs=5M count=100
     LOOP=$(losetup -f)
@@ -40,7 +54,9 @@ else
     LOOP=$(losetup -f)
     losetup $LOOP /opt/data/kolla/ceph-osd0-d.img
     parted $LOOP -s -- mklabel gpt mkpart KOLLA_CEPH_OSD_BOOTSTRAP_BS_OSD0_D 1 -1
+else
+    echo "Unknown type" >&2
+    exit 1
 fi
 
 partprobe
-
